combine send-block-kit-message & send-custom-message
