+++ components/slack/actions/send-custom-message/send-custom-message.mjs
@@ -1,7 +1,9 @@
 import common from "../common/send-message.mjs";
+import buildBlocks from "../common/build-blocks.mjs";
 
 export default {
   ...common,
+  ...buildBlocks,
   key: "slack-send-custom-message",
   name: "Send a Custom Message",
   description: "Customize advanced setttings and send a message to a channel, group or user. See [postMessage](https://api.slack.com/methods/chat.postMessage) or [scheduleMessage](https://api.slack.com/methods/chat.scheduleMessage) docs here",
@@ -20,6 +22,8 @@ export default {
         common.props.slack,
         "text",
       ],
+      description: "If you're using `blocks`, this is used as a fallback string to display in notifications. If you aren't, this is the main body text of the message. It can be formatted as plain text, or with mrkdwn.",
+      optional: true,
     },
     mrkdwn: {
       propDefinition: [
@@ -37,12 +41,6 @@ export default {
       propDefinition: [
         common.props.slack,
         "parse",
-      ],
-    },
-    blocks: {
-      propDefinition: [
-        common.props.slack,
-        "blocks",
       ],
     },
     link_names: {
@@ -66,7 +64,23 @@ export default {
         }),
       ],
       description: "Provide another message's `ts` value to make this message a reply (e.g., if triggering on new Slack messages, enter `{{event.ts}}`). Avoid using a reply's `ts` value; use its parent instead.",
+      optional: true,
     },
     ...common.props,
+    ...buildBlocks.props,
+  },
+  methods: {
+    ...common.methods,
+    ...buildBlocks.methods,
+    async getGeneratedBlocks() {
+      return await buildBlocks.run.call(this);  // call buildBlocks.run with the current context
+    },
+  },
+  async run({ $ }) {
+    this.blocks = await this.getGeneratedBlocks();  // set the blocks prop for common.run to use
+    const resp = await common.run.call(this, {
+      $,
+    });  // call common.run with the current context
+    return resp;
   },
 };
